+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adds `/whatisthisitem offhand`
 - `/wiki` autocomplete is now cached locally to improve performance
 - Now prompts user when `H` is pressed but nothing is being aimed at
+- Added configuration options for reach distance and whether fluid is allowed
 
 ## Fixes